+++ pkg/security/tests/cmdwrapper.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-//go:build linux && (functionaltests || stresstests)
+//go:build functionaltests || stresstests
 
 // Package tests holds tests related files
 package tests
