+++ tasks/security_agent.py
@@ -839,14 +839,16 @@ def kitchen_prepare(ctx, windows=is_windows, skip_linters=False):
         shutil.rmtree(KITCHEN_ARTIFACT_DIR)
 
     out_binary = "testsuite"
+    race = True
     if windows:
         out_binary = "testsuite.exe"
+        race = False
 
     testsuite_out_path = os.path.join(KITCHEN_ARTIFACT_DIR, "tests", out_binary)
     build_functional_tests(
         ctx,
         bundle_ebpf=False,
-        race=True,
+        race=race,
         debug=True,
         output=testsuite_out_path,
         skip_linters=skip_linters,
